Extended Practice - Final
Extended functionallity in practice 2 (Adoption of the new backing service made in practice 3) Also, created a new service, get petient by code, to demonstrate the new change. Also, te model was polished, the code parameter is not required and does not appear in swagger since is internally managed and generated.